Update server if #83

+++ ShiftManager.Communication.Common/CurrentTimeProvider.cs
@@ -7,5 +7,5 @@ namespace ShiftManager.Communication
     DateTime CurrentTime { get; }
   }
 
-  public class CurrentTimeProvider : ICurrentTimeProvider { public DateTime CurrentTime => DateTime.Today + TimeSpan.FromMinutes(DateTime.Now.TimeOfDay.TotalMinutes); } //分未満を削除する
+  public class CurrentTimeProvider : ICurrentTimeProvider { public DateTime CurrentTime => DateTime.Today + TimeSpan.FromMinutes((int)DateTime.Now.TimeOfDay.TotalMinutes); } //分未満を削除する
 }
